moved Code generated by protoc-gen-go out of internal so it can be referenced/used
