+++ tensorflow/go/.gitignore
@@ -1,2 +1,2 @@
 # .pb.go files generated by generate.sh
-internal/proto/*
+pb/proto/*
